requirements.txt
# Core AI APIs
google-genai>=1.0.0
anthropic>=0.40.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Databases
neo4j[async]>=5.25.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Vector store
chromadb>=0.5.0

# Caching
redis>=5.0.0

# Scheduling
apscheduler>=3.10.0

# HTTP client
aiohttp>=3.10.0

# Embeddings (used by ChromaDB default)
sentence-transformers>=3.0.0
